Initial commit with semi-working implementation
A fake click event is used to trigger the file input, which calls `webView(_:runOpenPanelWith:initiatedByFrame:completionHandler:)`. Calling `fileinput.click()` from JavaScript did not work, but sending a fake click event to the WKWebView does.

Currently the issue of file access from the sandbox is not being addressed; we are simply using dummy data that exists inside the preview extension's bundle.

Alternate approaches tried, which may eventually be useful if we run into sandboxing issues:
- Fake drag event with window.ondragstart/ondrop (more steps involved in sending the fake event; did not successfully get the web view to deliver the drag)
- HTTP server running in the app (though it appears a preview extension is not allowed to listen for incoming connections - but maybe an XPC service bundled in the preview extension would be able to?)